Refactor curl code a bit, to be more in line with all the other code.
